correcao dos exercicios 1 a 5 para inteiro

+++ exercicios.py
@@ -11,18 +11,18 @@
 # print(numero % 5)
 
 # 3. Desenvolva um programa que multiplique dois números fornecidos pelo usuário e mostre o resultado.
-# n1 = float(input("Digite o primeiro número: "))
-# n2 = float(input("DIgite o segundo número: "))
+# n1 = int(input("Digite o primeiro número: "))
+# n2 = int(input("DIgite o segundo número: "))
 # multiplicacao = n1 * n2
 # print(multiplicacao)
 
 # 4. Faça um programa que peça dois números inteiros e imprima a divisão inteira do primeiro pelo segundo.
-# n1 = float(input("Digite o primeiro número: "))
-# n2 = float(input("DIgite o segundo número: "))
+# n1 = int(input("Digite o primeiro número: "))
+# n2 = int(input("DIgite o segundo número: "))
 # print(n1 // n2)
 
 # 5. Escreva um programa que calcule o quadrado de um número fornecido pelo usuário.
-numero = float(input("Digite um número: "))
+numero = int(input("Digite um número: "))
 quadrado = numero ** 2
 print(quadrado)
